src/cloud_functions/data_scheduler/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.18.4
requests==2.31.0
